code cleaning up; minor improvement
